# openapi.json
{"openapi":"3.1.0","info":{"title":"TeremFlow","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/logout":{"post":{"tags":["auth"],"summary":"Logout","operationId":"logout_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/me":{"get":{"tags":["auth"],"summary":"Me","operationId":"me_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}}}}},"/cases/":{"get":{"tags":["cases"],"summary":"List Cases","operationId":"list_cases_cases__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/CaseOut"},"type":"array","title":"Response List Cases Cases  Get"}}}}}},"post":{"tags":["cases"],"summary":"Create Case","operationId":"create_case_cases__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaseCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaseOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases/{case_id}":{"get":{"tags":["cases"],"summary":"Get Case","operationId":"get_case_cases__case_id__get","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"integer","title":"Case Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaseOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases/{case_id}/status":{"patch":{"tags":["cases"],"summary":"Update Case Status","operationId":"update_case_status_cases__case_id__status_patch","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"integer","title":"Case Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaseUpdateStatus"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaseOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases/{case_id}/expenses/":{"get":{"tags":["expenses"],"summary":"List Expenses","operationId":"list_expenses_cases__case_id__expenses__get","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"integer","title":"Case Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ExpenseOut"},"title":"Response List Expenses Cases  Case Id  Expenses  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["expenses"],"summary":"Add Expense","operationId":"add_expense_cases__case_id__expenses__post","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"integer","title":"Case Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExpenseCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ExpenseOut"},"title":"Response Add Expense Cases  Case Id  Expenses  Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases/{case_id}/retainer/accruals":{"get":{"tags":["retainer"],"summary":"List Accruals","operationId":"list_accruals_cases__case_id__retainer_accruals_get","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"integer","title":"Case Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RetainerAccrualOut"},"title":"Response List Accruals Cases  Case Id  Retainer Accruals Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases/{case_id}/retainer/payments":{"get":{"tags":["retainer"],"summary":"List Payments","operationId":"list_payments_cases__case_id__retainer_payments_get","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"integer","title":"Case Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RetainerPaymentOut"},"title":"Response List Payments Cases  Case Id  Retainer Payments Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["retainer"],"summary":"Add Payment","operationId":"add_payment_cases__case_id__retainer_payments_post","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"integer","title":"Case Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RetainerPaymentCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RetainerPaymentOut"},"title":"Response Add Payment Cases  Case Id  Retainer Payments Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases/{case_id}/retainer/summary":{"get":{"tags":["retainer"],"summary":"Summary","operationId":"summary_cases__case_id__retainer_summary_get","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"integer","title":"Case Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RetainerSummary"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases/{case_id}/fees/":{"get":{"tags":["fees"],"summary":"List Fee Events","operationId":"list_fee_events_cases__case_id__fees__get","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"integer","title":"Case Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FeeEventOut"},"title":"Response List Fee Events Cases  Case Id  Fees  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["fees"],"summary":"Add Fee Event","operationId":"add_fee_event_cases__case_id__fees__post","parameters":[{"name":"case_id","in":"path","required":true,"schema":{"type":"integer","title":"Case Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeeEventCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeeEventOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/analytics/overview":{"get":{"tags":["analytics"],"summary":"Overview","operationId":"overview_analytics_overview_get","parameters":[{"name":"start_date","in":"query","required":true,"schema":{"type":"string","format":"date","title":"Start Date"}},{"name":"end_date","in":"query","required":true,"schema":{"type":"string","format":"date","title":"End Date"}},{"name":"case_type","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/CaseType"},{"type":"null"}],"title":"Case Type"}},{"name":"payer_status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Payer Status"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalyticsOverviewResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/import/excel":{"post":{"tags":["import"],"summary":"Import Excel","operationId":"import_excel_import_excel_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_import_excel_import_excel_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/notifications/":{"get":{"tags":["notifications"],"summary":"List Notifications","operationId":"list_notifications_notifications__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/NotificationOut"},"type":"array","title":"Response List Notifications Notifications  Get"}}}}}}},"/notifications/{notification_id}/read":{"post":{"tags":["notifications"],"summary":"Mark Read","operationId":"mark_read_notifications__notification_id__read_post","parameters":[{"name":"notification_id","in":"path","required":true,"schema":{"type":"integer","title":"Notification Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tasks/daily":{"post":{"tags":["tasks"],"summary":"Daily Tasks","operationId":"daily_tasks_tasks_daily_post","parameters":[{"name":"x-tasks-token","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Tasks-Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AnalyticsOverviewResponse":{"properties":{"total_expenses_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Total Expenses Ils Gross"},"total_on_deductible_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Total On Deductible Ils Gross"},"total_on_insurer_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Total On Insurer Ils Gross"},"average_expenses_per_case_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Average Expenses Per Case Ils Gross"},"cases_switched_to_insurer_count":{"type":"integer","title":"Cases Switched To Insurer Count"},"aggregate_remaining_deductible_open_cases_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Aggregate Remaining Deductible Open Cases Ils Gross"},"expenses_by_case":{"items":{"$ref":"#/components/schemas/ExpensesByCaseRow"},"type":"array","title":"Expenses By Case"},"expense_split":{"additionalProperties":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"},"type":"object","title":"Expense Split"},"court_cases_end_stage_distribution":{"items":{"$ref":"#/components/schemas/StageDistributionRow"},"type":"array","title":"Court Cases End Stage Distribution"},"monthly":{"items":{"$ref":"#/components/schemas/TimeSeriesPoint"},"type":"array","title":"Monthly"},"quarterly":{"items":{"$ref":"#/components/schemas/TimeSeriesPoint"},"type":"array","title":"Quarterly"},"yearly":{"items":{"$ref":"#/components/schemas/TimeSeriesPoint"},"type":"array","title":"Yearly"}},"type":"object","required":["total_expenses_ils_gross","total_on_deductible_ils_gross","total_on_insurer_ils_gross","average_expenses_per_case_ils_gross","cases_switched_to_insurer_count","aggregate_remaining_deductible_open_cases_ils_gross","expenses_by_case","expense_split","court_cases_end_stage_distribution","monthly","quarterly","yearly"],"title":"AnalyticsOverviewResponse"},"Body_import_excel_import_excel_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_import_excel_import_excel_post"},"CaseCreate":{"properties":{"case_reference":{"type":"string","maxLength":120,"minLength":2,"title":"Case Reference"},"case_type":{"$ref":"#/components/schemas/CaseType"},"open_date":{"type":"string","format":"date","title":"Open Date"},"deductible_usd":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"},{"type":"null"}],"title":"Deductible Usd"},"deductible_ils_gross":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"},{"type":"null"}],"title":"Deductible Ils Gross"}},"type":"object","required":["case_reference","case_type","open_date"],"title":"CaseCreate"},"CaseOut":{"properties":{"id":{"type":"integer","title":"Id"},"case_reference":{"type":"string","title":"Case Reference"},"case_type":{"$ref":"#/components/schemas/CaseType"},"status":{"$ref":"#/components/schemas/CaseStatus"},"open_date":{"type":"string","format":"date","title":"Open Date"},"deductible_usd":{"anyOf":[{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"},{"type":"null"}],"title":"Deductible Usd"},"fx_rate_usd_ils":{"anyOf":[{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"},{"type":"null"}],"title":"Fx Rate Usd Ils"},"fx_date_used":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Fx Date Used"},"fx_source":{"type":"string","title":"Fx Source"},"deductible_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Deductible Ils Gross"},"insurer_started":{"type":"boolean","title":"Insurer Started"},"insurer_start_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Insurer Start Date"},"deductible_remaining_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Deductible Remaining Ils Gross"}},"type":"object","required":["id","case_reference","case_type","status","open_date","deductible_usd","fx_rate_usd_ils","fx_date_used","fx_source","deductible_ils_gross","insurer_started","insurer_start_date","deductible_remaining_ils_gross"],"title":"CaseOut"},"CaseStatus":{"type":"string","enum":["OPEN","CLOSED"],"title":"CaseStatus"},"CaseType":{"type":"string","enum":["COURT","DEMAND_LETTER","SMALL_CLAIMS"],"title":"CaseType"},"CaseUpdateStatus":{"properties":{"status":{"$ref":"#/components/schemas/CaseStatus"}},"type":"object","required":["status"],"title":"CaseUpdateStatus"},"ExpenseCategory":{"type":"string","enum":["ATTORNEY_FEE","EXPERT","MEDICAL_INFO","INVESTIGATOR","FEES","OTHER"],"title":"ExpenseCategory"},"ExpenseCreate":{"properties":{"supplier_name":{"type":"string","maxLength":120,"minLength":2,"title":"Supplier Name"},"amount_ils_gross":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"}],"title":"Amount Ils Gross"},"service_description":{"type":"string","minLength":1,"title":"Service Description"},"demand_received_date":{"type":"string","format":"date","title":"Demand Received Date"},"expense_date":{"type":"string","format":"date","title":"Expense Date"},"category":{"$ref":"#/components/schemas/ExpenseCategory"},"payer":{"anyOf":[{"$ref":"#/components/schemas/ExpensePayer"},{"type":"null"}]},"attachment_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Attachment Url"}},"type":"object","required":["supplier_name","amount_ils_gross","service_description","demand_received_date","category"],"title":"ExpenseCreate"},"ExpenseOut":{"properties":{"id":{"type":"integer","title":"Id"},"case_id":{"type":"integer","title":"Case Id"},"supplier_name":{"type":"string","title":"Supplier Name"},"amount_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Amount Ils Gross"},"service_description":{"type":"string","title":"Service Description"},"demand_received_date":{"type":"string","format":"date","title":"Demand Received Date"},"expense_date":{"type":"string","format":"date","title":"Expense Date"},"category":{"$ref":"#/components/schemas/ExpenseCategory"},"payer":{"$ref":"#/components/schemas/ExpensePayer"},"attachment_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Attachment Url"},"split_group_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Split Group Id"},"is_split_part":{"type":"boolean","title":"Is Split Part"}},"type":"object","required":["id","case_id","supplier_name","amount_ils_gross","service_description","demand_received_date","expense_date","category","payer","attachment_url","split_group_id","is_split_part"],"title":"ExpenseOut"},"ExpensePayer":{"type":"string","enum":["CLIENT_DEDUCTIBLE","INSURER"],"title":"ExpensePayer"},"ExpensesByCaseRow":{"properties":{"case_id":{"type":"integer","title":"Case Id"},"case_reference":{"type":"string","title":"Case Reference"},"case_type":{"$ref":"#/components/schemas/CaseType"},"status":{"$ref":"#/components/schemas/CaseStatus"},"payer_status":{"type":"string","title":"Payer Status"},"total_expenses_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Total Expenses Ils Gross"},"attorney_fees_expenses_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Attorney Fees Expenses Ils Gross"},"other_expenses_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Other Expenses Ils Gross"},"deductible_remaining_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Deductible Remaining Ils Gross"}},"type":"object","required":["case_id","case_reference","case_type","status","payer_status","total_expenses_ils_gross","attorney_fees_expenses_ils_gross","other_expenses_ils_gross","deductible_remaining_ils_gross"],"title":"ExpensesByCaseRow"},"FeeEventCreate":{"properties":{"event_type":{"$ref":"#/components/schemas/FeeEventType"},"event_date":{"type":"string","format":"date","title":"Event Date"},"quantity":{"type":"integer","minimum":1.0,"title":"Quantity","default":1},"amount_override_ils_gross":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"},{"type":"null"}],"title":"Amount Override Ils Gross"}},"type":"object","required":["event_type","event_date"],"title":"FeeEventCreate"},"FeeEventOut":{"properties":{"id":{"type":"integer","title":"Id"},"event_type":{"$ref":"#/components/schemas/FeeEventType"},"event_date":{"type":"string","format":"date","title":"Event Date"},"quantity":{"type":"integer","title":"Quantity"},"amount_override_ils_gross":{"anyOf":[{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"},{"type":"null"}],"title":"Amount Override Ils Gross"},"computed_amount_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Computed Amount Ils Gross"},"amount_covered_by_credit_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Amount Covered By Credit Ils Gross"},"amount_due_cash_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Amount Due Cash Ils Gross"}},"type":"object","required":["id","event_type","event_date","quantity","amount_override_ils_gross","computed_amount_ils_gross","amount_covered_by_credit_ils_gross","amount_due_cash_ils_gross"],"title":"FeeEventOut"},"FeeEventType":{"type":"string","enum":["COURT_STAGE_1_DEFENSE","COURT_STAGE_2_DAMAGES","COURT_STAGE_3_EVIDENCE","COURT_STAGE_4_PROOFS","COURT_STAGE_5_SUMMARIES","AMENDED_DEFENSE_PARTIAL","AMENDED_DEFENSE_FULL","THIRD_PARTY_NOTICE","ADDITIONAL_PROOF_HEARING","DEMAND_FIX","DEMAND_HOURLY","SMALL_CLAIMS_MANUAL"],"title":"FeeEventType"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"username":{"type":"string","maxLength":50,"minLength":2,"title":"Username"},"password":{"type":"string","maxLength":200,"minLength":7,"title":"Password"}},"type":"object","required":["username","password"],"title":"LoginRequest"},"NotificationOut":{"properties":{"id":{"type":"integer","title":"Id"},"case_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Case Id"},"type":{"$ref":"#/components/schemas/NotificationType"},"title":{"type":"string","title":"Title"},"message":{"type":"string","title":"Message"},"severity":{"type":"string","title":"Severity"},"is_read":{"type":"boolean","title":"Is Read"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","case_id","type","title","message","severity","is_read","created_at"],"title":"NotificationOut"},"NotificationType":{"type":"string","enum":["DEDUCTIBLE_NEAR_EXHAUSTION","INSURER_STARTED_PAYING","RETAINER_DUE_SOON","RETAINER_OVERDUE"],"title":"NotificationType"},"RetainerAccrualOut":{"properties":{"id":{"type":"integer","title":"Id"},"accrual_month":{"type":"string","format":"date","title":"Accrual Month"},"invoice_date":{"type":"string","format":"date","title":"Invoice Date"},"due_date":{"type":"string","format":"date","title":"Due Date"},"amount_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Amount Ils Gross"},"is_paid":{"type":"boolean","title":"Is Paid"}},"type":"object","required":["id","accrual_month","invoice_date","due_date","amount_ils_gross","is_paid"],"title":"RetainerAccrualOut"},"RetainerPaymentCreate":{"properties":{"payment_date":{"type":"string","format":"date","title":"Payment Date"},"amount_ils_gross":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"}],"title":"Amount Ils Gross"}},"type":"object","required":["payment_date","amount_ils_gross"],"title":"RetainerPaymentCreate"},"RetainerPaymentOut":{"properties":{"id":{"type":"integer","title":"Id"},"payment_date":{"type":"string","format":"date","title":"Payment Date"},"amount_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Amount Ils Gross"}},"type":"object","required":["id","payment_date","amount_ils_gross"],"title":"RetainerPaymentOut"},"RetainerSummary":{"properties":{"retainer_accrued_total_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Retainer Accrued Total Ils Gross"},"retainer_paid_total_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Retainer Paid Total Ils Gross"},"retainer_applied_to_fees_total_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Retainer Applied To Fees Total Ils Gross"},"retainer_credit_balance_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Retainer Credit Balance Ils Gross"},"fees_due_total_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Fees Due Total Ils Gross"}},"type":"object","required":["retainer_accrued_total_ils_gross","retainer_paid_total_ils_gross","retainer_applied_to_fees_total_ils_gross","retainer_credit_balance_ils_gross","fees_due_total_ils_gross"],"title":"RetainerSummary"},"StageDistributionRow":{"properties":{"stage":{"type":"integer","title":"Stage"},"count":{"type":"integer","title":"Count"}},"type":"object","required":["stage","count"],"title":"StageDistributionRow"},"TimeSeriesPoint":{"properties":{"period":{"type":"string","title":"Period"},"total_expenses_ils_gross":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Total Expenses Ils Gross"}},"type":"object","required":["period","total_expenses_ils_gross"],"title":"TimeSeriesPoint"},"UserOut":{"properties":{"id":{"type":"integer","title":"Id"},"username":{"type":"string","title":"Username"},"role":{"type":"string","title":"Role"}},"type":"object","required":["id","username","role"],"title":"UserOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}